drivers: platform: maxim: max32660: Add SPI delays
Implement support for the SPI delays on the MAX32660. Since the registers
for the cs_delay_first and cs_delay_last are 8 bits wide and the value is
measured in peripheral clock ticks (which has a frequency of 60 MHz), the
maximum delay which may be set is 4 us. Anything higher than that will
result in using the delays from the previous transfer.

Also removed some extra lines.

Signed-off-by: Ciprian Regus <ciprian.regus@analog.com>
(cherry picked from commit cf95639872857a635d0b1cd2050b40734548375c)