add MDIO API
MDIO is used to control networking PHYs from MACs.
An API for the interface is needed in no-OS and this commit
defines an API that allows the user to create an MDIO device
desriptor per MDIO slave.

It is possible to select between multiple implementations using
ops.

Once the MDIO device descriptor is initialized, it is possible
to read/write slave registers via MDIO interface using either
clause 22 or clause 45 frame format.

Signed-off-by: Darius Berghe <darius.berghe@analog.com>