iio: iiod: Optimized network buffer transfers
Currently, reading buffered data from a device causes the IIOD to send it
through the backend link as soon as there are samples available. Because
in most cases the IIOD_RW_BUF state is reached more often than the
device buffered is completely filled, there will be underrun, resulting
in transmitting just a few samples at a time. This is not an issue for
the serial backend (the overhead is the same regardless of the samples
number). On the network backend however, this results in transmitting
multiple minimum sized frames (i.e 4 bytes of sample data will be
encoded into a 64 byte frame), which is inefficient.

Fix this issue by waiting for min(buffer_size, n_samples) samples to be
stored in the buffer before a transmission is issued.

Signed-off-by: Ciprian Regus <ciprian.regus@analog.com>
(cherry picked from commit b4fa0329f08bb70a4c787194ba7ad694e7e12785)