iio_app: Add default trigger id to iio devices
In some cases we want to assign a trigger to a device
at init time. This implementation adds default_trigger_id
parameter to iio_app_device structure such that a trigger
can be assigned to a device at init time.
The trigger id should be of the form "trigger0", "trigger1",
etc.

Signed-off-by: Ramona Bolboaca <ramona.bolboaca@analog.com>