iio: Lwip tcp sockets support
Add a network backend support for IIOD using the lwip TCP sockets
implementation. The NO_OS_LWIP_NETWORKING macro is defined to enable the
use of the lwip layer. This is done in order to avoid breaking the
current implementation, which interprets what networking is based on the
platform used.

Signed-off-by: Ciprian Regus <ciprian.regus@analog.com>