drivers: crypto: maxq1065: Fix function signature
The maxq1065_first_bytes_spi() function should receive 2 parameters
instead of 3. Fix this.

The correct way to return the data from the SPI RX functions is to just
copy into the destination buffer, without swapping the bits (because of
the LSB/MSB conversion).

Both MXQ_Ping() and MXQ_GetStatus() commands are working.

Signed-off-by: Ciprian Regus <ciprian.regus@analog.com>