drivers:platform:maxim: Fix UART TX
The SDK implemententation for revB MXC_UART_Write() has asynchronous
behavior (meaning that if the TX FIFO is full, it will return the number
of bytes written). This requires additional handling in the application
code. Apply the 75cdb79("drivers:platform:max78000: Fix blocking UART Tx")
for all the other Maxim MCUs, which force the max_uart_write() function
to block until all the data is transmitted.

Note that the fix is not required for chips that use the revA version of
the driver, since that implementation is already blocking. In such case,
the additional handling will make sure the current behavior is kept if
anything changes in the SDK driver.

Signed-off-by: Ciprian Regus <ciprian.regus@analog.com>