drivers:digital-io:max14906: Fix input setting
All the MAX14909's channels are set to high impedance mode during driver
init. Hardware wise, this is done by setting a channel to input,
configuring the output driver as push-pull.
The issue is that later, when a channel is configured as input, the
push-pull setting is not changed, this the channel continues to be in
high impedance. Fix this by setting the driver in high-side mode.

Signed-off-by: Ciprian Regus <ciprian.regus@analog.com>