tools/scripts: Download hardware from artifactory
This script has the role to download latest hardware files from artifactory.
If the hardware file of the project is missing from the lates timestamp(build)
then it will check for older hardware file in previouse builds of the hdl.

For accessing the server path where the new hardware files are located and
for the path to platform_environment.sh scripts it will be needed a configuration.txt
file where this values are saved, due to privacy issues.

The new hardware files will be downloaded just for the xilinx platform and this
is done by running the download_files.py script with propper parametes. This
script can be run separately.

Signed-off-by: Andreea Andrisan <Andreea.Andrisan@analog.com>