add no-OS project name to iio context description
Add no-OS project name to the iio context description.

libiio v0.25 now handles the context description correctly with
the serial backend by displaying it as-is instead of replacing
it with other information.

This is useful to iio apps that need to know what firmware is
running on the board (project name, git tag, revision).

Signed-off-by: Darius Berghe <darius.berghe@analog.com>