drivers:adc-dac:ad74413r: Make trigger synchronous
With the current (asynchronous) trigger setting, running the AD74413R at
4.8 KSPS will result in loosing samples, since the IIOD state machine
is not able to process the async trigger function every 200 us. Fix
this issue by making the trigger synchronous. This way, a sample will be
read from the device on each data ready interrupt.

Move the iio_trigger struct to be project specific.

Signed-off-by: Ciprian Regus <ciprian.regus@analog.com>