drivers:adc-dac:ad74413r: Add optional reset GPIO
For the AD74413R to be brought out of reset, the RESET pin must be
pulled high. Currently, this signal must either be tied high via a
hardware connection or be controlled via a GPIO (independent of the
driver). Implement this handling in the driver by letting the user to
optionally provide the reset GPIO parameter.

Signed-off-by: Ciprian Regus <ciprian.regus@analog.com>