iio: Add support for delta angle and delta velocity
Add support for delta angle and delta velocity to be
to have the same types as in Linux IIO implementation.

The delta angle is defined as a piece-wise integration of angular
velocity data. The delta angle represents the amount of
angular displacement between two consecutive measurements and it
is measured in radians.

The delta velocity is defined as a piece-wise integration of
acceleration data. The delta velocity represents the linear velocity
change between two consecutive measurements and it
is measured in m / s (meters per second).

Signed-off-by: Ramona Gradinariu <ramona.gradinariu@analog.com>