tools: scripts: migrate the release scripts to ci
The release scripts are implemented and used only based on Azure
Pipelines procedures. Since the `tools/scripts` folder contain
scripts/makefiles that can be used locally and are related to the
platform/projects build, move the release scripts to a more relevant
location, under `ci/release`. In this manner we keep things cleaner.

Signed-off-by: Antoniu Miclaus <antoniu.miclaus@analog.com>