Allow for one time init of the lwip stack
In case DHCP is attempted, initializing lwip takes a long time (~10s).
Currently, this is done each time iio_app_init() is called. Add an
option to only configure the networking stack one time, since there
might not be a reason to restart the process if multiple
iio_app_init()/iio_app_remove() sequences are called, with the intent of
updating the IIO devices.

Signed-off-by: Ciprian Regus <ciprian.regus@analog.com>