network:lwip_sockets: Fix pbuf ref counting
When adding a new packet buffer to a socket's pbuf list, the new pbuf
shouldn't have it's reference count increased, since it should only be
referenced by the previous element in the list. By using pbuf_chain(),
the reference count is increased, which is wrong, since the pbuf will
not be used outside of list reading mechanism. This results in a memory
leak, since the pbuf will not actually be freed.

pbuf_cat() does the concatenation without increasing the reference
count, thus this function should be used instead.

Fixes: 357434e3b89a ("network: lwip_raw_socket: Add generic lwip layer")
Signed-off-by: Ciprian Regus <ciprian.regus@analog.com>