build_projects.yml: add hdl_branch parameter
Instead of going on default case, parametrize the hdl_branch used for
projects builds using a an Azure environment variable alongside LOGS_DIR
and RELEASE_DIR passed to the build_projects.py script. In this manner
we keep under control the versions of the hdl used for building projects
and folder structure of the builds.

Signed-off-by: Antoniu Miclaus <antoniu.miclaus@analog.com>