drivers/axi_core/axi_pwmgen/axi_pwm: Use phase_ns init parameter
The setter for the axi_pwm uses the phase_ns member of the pwm_desc
struct, which was previously not initialised. 'pwm_desc' needs to be fully
initiliased using the info in 'param' struct.
This feature is needed for the SPIE offload trigger signal, which has
to be generated with a particular phase shift w.r.t. the ODR signal.

Signed-off-by: laurent-19 <laurentiu.popa@analog.com>
Signed-off-by: laurentiu_popa <Laurentiu.Popa@analog.com>