ci: doxygen: create .nojekyll file
Prior to pushing new documentation build on gh-pages, create .nojekyll
file in the root directory to avoid treating github pages as a Jekyll
site.

Sphinx generates documentation in a way that is different from Jekyll,
and we want to avoid any potential conflicts. In this specific case, the
files from the `_static` are ignored by Jekyll and we don't want that to
happen.

Signed-off-by: Antoniu Miclaus <antoniu.miclaus@analog.com>