update pid hysteresis and parameters
This commit fixes two things:
1) hysteresis was not being updated when the user would change it, it
was a constant 20 mV.
2) PID was made to be more aggressive in finding the target with the
guarantee (through simulation) that in 10 iterations the target was
found. The loop therefore ends after exactly 10 iterations.

Regarding point number 2, probably the iteration count can be squeezed
even lower to 5 or something like that, if processing time becomes a
concern. Keeping it to 10 to make sure the pid control does finish in
a single run.

Signed-off-by: Darius Berghe <darius.berghe@analog.com>