drivers: ad2s1210: Add excitation frequency set and get
Ad2s1210 provides an on-board programmable sinusoidal oscillator
that provides sine wave excitation for resolvers. which is
programmable to a number of standard frequencies between 2 kHz
and 20 kHz.

Allow to set and get the excitation frequency so that users of
this driver can provide a output channel for this oscillator.

Signed-off-by: Axel Haslam <ahaslam@baylibre.com>