tools: scripts: handling STM32 random failure
For the build of STM32, in some cases a random failure appears,
to handle this case, a new run of the failed project is made
right after the first fail. The rerun is made just one more time
in order to check if the first fail was a random one or the error
persists in the project.
If the new building status turns to be a good one then the previous
build will be interpreted as a random one and in the end the build
as a valid one.
If the new rerun exists also with error status this will be interpreted
as a persistent error not a random one.

In case a project is build twice it is mentioned in the log message.
This workaround is just for the STM32 projects, the other platforms
do not replicate this behavior.

Signed-off-by: Andreea Andrisan <Andreea.Andrisan@analog.com>
(cherry picked from commit 0b9e6ab4bd09a8275f47821a4d8f68df9abb05c7)