drivers: accel: adxl367: update self test delay
The wait times in the self test procedure, according to the datasheet
are 4 / ODR (current set value). Moreover, the current value set as
default delay (80ms) does not correspond to the lowest ODR value, as
stated in the code comments.

Update the self test procedure by using the delay corresponding to the
current ODR value that is set.

Fixes: 2e4f0c3 ("drivers:accel: Create ADXL367 Driver")
Signed-off-by: Antoniu Miclaus <antoniu.miclaus@analog.com>