drivers: ad2s1210: Use mask to select conversion data
Currently "single_conversion" returns either a position
or a velocity of a sample, but not both at the same
time.

Pass a bitmask of active channels as parameter so that
the user can either request position, velocity or both.

This will allow to graph position and velocity
of the same sample.

Signed-off-by: Axel Haslam <ahaslam@baylibre.com>