drivers: adc-dac: ad74413r: Fix diag rate setting
The ad74413r_reg_update() does the field prep. When setting the sampling
rate for a diagnostics channel, the bitfield is shifted twice, which is
wrong.

Also, there is already a function that reads a register value and
converts it into a rejection value. So, this can be used instead of
directly reading the register.

Signed-off-by: Ciprian Regus <ciprian.regus@analog.com>