drivers: dac: ad5791: lin_comp error handling
The `enum ad5791_lin_comp_select` contains values for both ad5781 and
ad5791 parts with respect to the voltage spans. Altough, the only value
that is valid for both parts is `AD5781_SPAN_UPTO_10V`.

Update the error handling procedure such that incorrect values are not
allowed to be set, based on the device type used with the driver.

Signed-off-by: Antoniu Miclaus <antoniu.miclaus@analog.com>