platform:maxim: Remove the extra pointer free
For the Maxim platform, the extra field in the init param struct is not
used in the GPIO IRQ driver. Instead it's assigned to a pointer in the
descriptor structure, which allows the user to pass arbitrary data they
may need in this context. Freeing this memory may lead to undefined
behavior, since it may be statically allocated.

Signed-off-by: Ciprian Regus <ciprian.regus@analog.com>