ci: documentation: add extra sphinx checks
When a README.rst file is created/modified for a specific driver or
project, the documentation CI checks if also the links to the table of
contents are properly handled.

If no link is found under drivers_doc.rst or projects_doc.rst for that
specific driver/project, the documentation fails.

If no .rst file is found for including the driver/project README.rst
specific file, the documentation build fails.

Signed-off-by: Antoniu Miclaus <antoniu.miclaus@analog.com>