tools: scripts: generic_variables: handle maxim
If a TARGET is specified as parameter to the make command and has the
`max` prefix, then assume that the PLATFORM=maxim.

This patch allows building maxim platform projects without specifying
explicitly the platform if there is a maxim TARGET set.

Example:
`make PLATFORM=maxim TARGET=max32650`
can be called now simply
`make TARGET=max32650`

A similar approach is already implemented for the xilinx projects where
if a user sets the HARDWARE variable, the PLATFORM is set automatically
to `xilinx`.

Signed-off-by: Antoniu Miclaus <antoniu.miclaus@analog.com>