change the way adc channel 1 is configured
I encountered a saturation issue on one of the boards
when dividing the input by 2 and comparing it with 1.22V reference
(which should give good precision up to 2.44V). I should mention that
2.3...2.4V is the maximum read values in a real-world scenario for
RX_DET.

What I'm doing now is that I divide the input by 5 and compare it
with the 1.22 / 2 = 0.61 V but even though this essentially is more
or less the same thing (slightly less accurate, somehow I magically
don't have saturation issues by doing so.

Signed-off-by: Darius Berghe <darius.berghe@analog.com>