condition the vs code project configuration
This is currently only supported on stm32, so limit it to that
platform. This fixes some issues with building for xilinx platform
which has a funky way of detecting the architecture (build/tmp/arch.txt)
which makes the builds fail. The failure is due to a which command
call without arguments and it all comes from vs code support.

Signed-off-by: Darius Berghe <darius.berghe@analog.com>