SPI: STM32: Dont enable/disable HAL on transfer
On SPI MODE 2, the first bit of data sent by the salve
is wrong.

This happens because the platform spi driver disables
the SPI HAL after each trasfer reulting on the clk line
beeing put back to its default floating state. It may
go to 0v or 1.8v depending on the gpio pull configuration.
The spi transactions level is 3.3v.

Because of the software controlled CS, the CS pin is active
when HAL is enabled for the next xfer creating an invalid
clk edge seen by the slave. The result is that the fist bit
of the data transfer is wrong.

Keep the HAL enabled and avoid enabling and disabling
on each trasfer.

Signed-off-by: Axel Haslam <ahaslam@baylibre.com>