projects: swiot1l: Use soft reset for the AD74413R
The reset signal for the AD74413R is routed through a digital isolator,
which samples the line by polling. Because of this, the output reset
signal has a variable jitter (20 us maximum), which causes the reset to
not be reliable.

Work around this by setting the reset GPIO high and using soft reset
instead.

Signed-off-by: Ciprian Regus <ciprian.regus@analog.com>