Add a generic DMA API
Add a DMA API, which is meant to handle scatter-gather memory to memory,
memory to peripheral and peripheral to memory transfers.

This layer abstracts a platform's DMA as one or more DMA controllers,
each having one or more DMA channels (which are not shared between
controllers). Each channel has a list of transfers, which by default are
cycled between using an end of transfer interrupt handler, that is also
implemented as part of the generic DMA layer. In case the interrupt
mechanism is different for a particular DMA IP, the handling has to be
done by accessing the list in order to enqueue or dequeue transfers.

Signed-off-by: Ciprian Regus <ciprian.regus@analog.com>