maxim: Set the IRQ action lists to NULL on remove
max_irq_register_callback() initializes a list of actions only if the
that list is NULL. However, after the init(), remove(), init() sequence,
the list is not set to NULL, even though it was freed and it has to be
allocated again in order to be used. As such, we end up accessing a
random memory location when registering a callback.

Fix this by setting the action lists to NULL in remove().

Signed-off-by: Ciprian Regus <ciprian.regus@analog.com>