doc: Add Makefile and repository to conf.py
Adds the default Sphinx Makefile with vars SOURCEDIR and BUILDDIR set.
It needs to exist because doctools needs to know the source and build paths.
Set values are consistent with the ones at documentation.sh.
Created it in the doc/sphinx/source path instead of doc/sphinx because it
would be the only file at the latter path.
Also, includes repository var to conf.py, used by doctools.

Signed-off-by: Jorge Marques <jorge.marques@analog.com>