tools: scripts: mbed: update build path
The MBED_OS_BUILD_DIRECTORY holds the build files that the mbed platform
generates. These files should be placed in the same manner as all the
other platform, respectively under the BUILD_DIR environment path.

This patch fixes the CI builds for the mbed platform and also makes sure
that all the build generated files are kept under the same path when
building manually a project under the `projects` folder.

Signed-off-by: Antoniu Miclaus <antoniu.miclaus@analog.com>