stm32: spi: make cs optional
Some chips (like the ad 400x series) dont have a
CS, and use the CNV pin (or the SDI) pins as CS.
Sometimes the CNV need to be held HIGH and not low
while data is read.

For these scenarios make the SPI CS optional so
that the driver will handle it before the spi
operation.

Signed-off-by: Axel Haslam <ahaslam@baylibre.com>