drivers: imu: adis: Rework scale implementation
Currently the scale implementation is not optimal: the arrays used
for scale which are defined in the chip specific drivers have a
bigger size than needed because device id is used.
This commit removes this problem by:
- adding chip specific API declaration to retrieve the scale,
instead of using scale arrays.
- implementing the chip specific APIs to retrieve the scale for
each existing adis driver
- using a device id with no offset in each chip specific driver
to define the scales for all supported devices.
All these changes lead to correct scale reading from adis driver
based on the device id, with driver-specific scale array size
equal to the number of supported devices.
Unit-tests are also updated to reflect this change.

Signed-off-by: Ramona Gradinariu <ramona.gradinariu@analog.com>