platform: maxim: max32690: Include gpio.h
Currently, the gpio.h header is not included in some header files which
use GPIO specific data types. This may or may not result in compilation
errors depending if a header which includes gpio.h is included before in
a specific source file.

Fix this by explicitly including gpio.h whenever necessary.

Signed-off-by: Ciprian Regus <ciprian.regus@analog.com>