ad40xx: fix spi_single_conversion function
If the SPI_ENGINE_OFFLOAD_EXAMPLE=0, it is trigger the spi single
conversion mode. Before this commit all the samples returned was 0. In
HDL the ad40xx_spi_sdo was set to 1'b1 value. Due to the limitation of
spi_engine implemented in NO-OS, it is needed 24 bits for the conversion, of
which I masked 20 bits. With the offload switch variable, the data width is
set, depending on the conversion mode that it is set through the
SPI_ENGINE_OFFLOAD_EXAMPLE parameter. The displayed signals are
represented in two's complement format.

Signed-off-by: Ioan-daniel Pop <Pop.Ioan-daniel@analog.com>