drivers: platform: maxim: Allow for I2C pins selection
Some microcontrollers have multiple pin mappings for the same I2C
interface. These configurations are not always defined in the SDK, which
makes it impossible to use them (unless the SDK definitions are
changed). Work around such cases by allowing the user to select which
signal should be used.

The MXC_I2C_Init() function may no longer be used, since it will always
select a set of default I2C pins for a specific interface.

As a result of the _max_i2c_pins_config() definition change, the eparam
variable is now unused, so we can remove it.

Signed-off-by: Ciprian Regus <ciprian.regus@analog.com>