maxim usb uart implementation
This is the implementation of the usb ops that use a USB CDC ACM
class for serial communication.

Note that the typical device id, baudrate, parity, stopbits settings
of a normal uart are ignored. The only required members of the init
param are the new ops and the irq_id (of the USB interrupt). There is
also a requirement that asynchronous_rx be true, because it is not
feasible to make this work without interrupts.

At the time of this commit, tested on the max32650 and max32690 only!
max32665 features USB peripheral as-well but it is untested.

Signed-off-by: Darius Berghe <darius.berghe@analog.com>