drivers: axi_core: axi_pwmgen: Add support for AXI PWM Generator v2 IP Core
This commit adds support for the AXI PWM Generator v2 IP Core. The main changes include
increasing the maximum number of channels to 16 and correcting register offsets for the
v2 implementation.

To maintain compatibility with the v1 implementation, a runtime check has been added to
choose the appropriate register configuration based on the hardware version.

Signed-off-by: Jonathan Nunes <jonathan.santos@analog.com>