tools: scripts: Set the path to LSCRIPT as absolute for Maxim
If the MAXIM_LIBRARIES variable is given as an argument for the make
command, the path to the user's home directory is not resolved to an
abosolute path. The linker cannot handle relative paths given to the -T
flag, and will generate an error:

/usr/lib/gcc/arm-none-eabi/12.2.1/../../../arm-none-eabi/bin/ld: cannot open linker script file

Fix this by replacing ~ with the path to the user's home directory. This
should only have an effect when compiling on Linux.

Signed-off-by: Ciprian Regus <ciprian.regus@analog.com>