projects: apard32690: Define IP related macros before generic.mk
In case we want to use a static IP, the NO_OS_IP, NO_OS_NETMASK and
NO_OS_GATEWAY variables have to be set before including generic.mk.
Otherwise, the macros with the same name won't be defined in lwip.mk,
thus the static IP settings will be ignored (resulting in trying to get
an IP over DHCP).

Fixes: b51e518 ("Allow for IP setting at compile time")
Reported-by: Brandon Hurst <brandon.hurst@analog.com>
Signed-off-by: Ciprian Regus <ciprian.regus@analog.com>