driver: adc-dac: ad74413r: Change conversion result type to signed
When the AD74413R's ADC is measuring current in the voltage output mode,
the result may be signed. A negative current indicates the current is
sourced by the device and a positive one indicates current is sinked.
Currently, since the decimal part of the result is declared as unsigned,
there is no way to represent current values in the (-1:0) mA range.

Fix this by declaring the decimal part as signed.

Signed-off-by: Ciprian Regus <ciprian.regus@analog.com>