tools: scripts: Delete .vscode on reset
After 86d407b("tools:scripts: Don't overwrite .vscode config if it already
exists") the vscode configuration files (contents of the .vscode
directory) are no longer overwritten on every make build. This may
lead to confusion if the build was accidentally done for a different
platform than the one intended.

Such a situation may occur if the user runs "make build" (without the
PLATFORM variable), resulting in generating the vscode configurations
for the STM32 platform. From this point there is no way to change the
target to another platform using the build system. Fix this by removing
the .vscode directory on "make reset".

Signed-off-by: Ciprian Regus <ciprian.regus@analog.com>