drivers: ad469x: Make number of channels configurable
Make the number of channels configurable so a project can configure
these for different parts or eval boards.

Because temperature is treated as an extra channel, we also add
accessor methods to check for the number of channels and if a
particular channel is the temperature channel. This helps make the
logic on project more clear to read.

Signed-off-by: Axel Haslam <ahaslam@baylibre.com>