drivers: ad469x: Add extended init config option
Allow for reduced complexity of project code by adding an optional config
flag to enable default sane settings for sequencer modes. This allows
to have the the device configured and ready to read samples after init
exits without error.

This option is off by default, and even when enabled, projects can
still choose to make their own config if desired.

For this:

- add option to enable this feature, so that its off by default for
  backwards compatibility.
- use basic ch/slot allocation: (ch0 -> slot0, ch1 -> slot1, ...)
- make temperature enable configurable on init
- enabling conversion mode before exit.

Signed-off-by: Axel Haslam <ahaslam@baylibre.com>