iio: Optimize IIO buffer transmission when using the network backend
Sending data over TCP is expensive since L4, L3 and L2 protocols each
add their own headers, resulting in a minimum ethernet frame size of 64
bytes. The current implementation of do_read_buff() favors UART traffic,
since trigger sampled data is sent as soon as it's available, thus the
transmission delay is reduced.

TCP is the most efficient when transfering larger chunks of data (as in
lower overhead). Because of that, wait for a whole IIO buffer to fill
before transmitting over the wire.

Signed-off-by: Ciprian Regus <ciprian.regus@analog.com>