projects: ad-acevsecrdset-sl: Added upstream PE detection functionality
This applies only to rev_D.
Added an interrupt on PE_error pin (P2.3).
Added self test functionality for PE upstream detection.
Added interrupts for opto_out1(P2.5) and opto_out2(P2.7) to check the
PE upstream availability. If the result of xor between them is zero than
PE not available.
When the upstream PE error triggers a debug error message is shown and the error is latching.
Hard reset of the device is needed.

Signed-off-by: Radu Etz <radu.etz@analog.com>
Signed-off-by: George Mois  <george.mois@analog.com>