drivers: imu: adis: Update read reg transfer
For 4 bytes reading, on tx add the read reg command instead of the page
update command.
This is a preparatory patch for adding support for adis1654x family.
For current devices which do not support paging the functionality is
not changed. For adis1654x this change is necessary to be able to
perform correctly the reading operation.

Signed-off-by: Ramona Gradinariu <ramona.gradinariu@analog.com>