platform: stm32: spi: fix dma callback
Currently there is no callback to undo the setup steps done dma xfer
is started. But also the callback passed as a parameter is not called when
dma xfer is done.

This is because dma driver uses the channel it allocates itself for the callback.
Use the channel allocated by the dma driver and assign a callback to it
that will free the memory and stop the dma/pwm that was started
on start.

Signed-off-by: Axel Haslam <ahaslam@baylibre.com>