projects: ad463x_fmcz: stm32: Enable dma for standard spi
Modify the ad463x_fmcz project to use dma for spi.

Eeach xfer 6 bytes are read, using DMA. Since CS and CNV pins
need to be coorinated for conversion then read, a regular DMA
transfer cannot be done.

Instead timers are used to generate the require signals in order,
and trigger the DMA rx when appropiate.

For this 3 timers are used:

1) "trigger" timer: toggles the CNV pin. master that starts the
                      sample sequence.

2) "CS" timer: slave to CNV timer toggles the CS pin (one shot)

3) "TX" timer: slave to CS timer, triggers dma writes to spi DR.
               set in one shot mode with 5 repetitions (total 6 bytes)

Signed-off-by: Axel Haslam <ahaslam@baylibre.com>