drivers: api: dma: dont call abort xfer on config xfer error
Abort xfer disables irq that is not registered yet on this error
path, and also calls platform xfer abort for a xfer that was not
started.

avoid calls to xfer abort and move all the sg list code to
the irq block since its use only for the irq case.

Signed-off-by: Axel Haslam <ahaslam@baylibre.com>