drivers: axi_core: Fix offload xfer size
spi_engine_offload_transfer should xfer number of samples,
not number of bytes.

Drivers are sometimes sending number of bytes, causing
a buffer over run. because the dma buffer is cyclic,
the overrun migh not be visible with a simple sine wave
or DC input as feed signal to the part.

The bytes that are transfered on dma per sample depend
on the DMA_DATA_WIDTH_SRC of the dma IP not on the data_width
bytes sent by the spi_engine on the spi bus.

Example:
for a 24 bit part (data_width of 3) if spi_engine_offload_transfer
asks for 100 samples, only 75 are done.

This is because 3 * 1 * 100 = 300 bytes,
but since the dma core is doing 4 byte xfers and the 4th byte
is counted too:
  300/4 = 75 total samples.

Fix this by using the acutal src size of the the dma ip to
calculate the number of bytes needed to complete n samples.

Signed-off-by: Axel Haslam <ahaslam@baylibre.com>