drivers: iio_ad463x: Avoid extra buffer allocation
The buffer size is such that it can hold data for all channels
as the driver will read always all channels.

In case of only one channel is active we dont need to allocate
a new buffer. Just use the existing buffer and shift channel data
replacing the deactivated channel.

Signed-off-by: Axel Haslam <ahaslam@baylibre.com>