stm32: Fix STM32CUBEIDE build (Linux and Windows)
generic.mk is linking $(ASM_SRCS), so, the startup file ends up being included
twice. This breaks the build of the project in STM32CUBEIDE.

Use a different name (e.g., STARTUP_FILE instead of ASM_SRCS) since the intend
in this case is to use this information only locally.

Signed-off-by: Dragos Bogdan <dragos.bogdan@analog.com>