drivers: power: max42500: Add the driver files for max42500
The max42500 is a power-system monitor with up to seven voltage monitor inputs and each input has programmable OV/UV thresholds. The source file contains the functions needed for the initialization of the device, reading the registers, writing to the registers, and removing the device. The header file contains the macros needed for configuration of the device.

Signed-off-by: Joshua Maniti <Joshua.Maniti@analog.com>