drivers: spi_engine: introduce error paths in spi_engine_offload_transfer()
While working on the AD7606 driver, an eval-board was found to be broken.
The failure message was hidden by a bunch of random data from memory, and
the part did not ever collect any samples.

This change updates the error paths in the spi_engine_offload_transfer()
function, so that (if an error occurs) it gets propagated correctly.

Signed-off-by: Alexandru Ardelean <aardelean@baylibre.com>