drivers: ad7606: rework channel ranges configuration
Especially for AD7606C-{16,18} we need to be able to specify the correct
register values for the ranges provided by IIO (and not just IIO).

The AD7606C-{16,18} channels (in SW mode) can be configured to be
- single-ended bipolar,
- single-ended unipolar
- differential bipolar

But, for IIO we also need to know the type of the channel (whether it's
one of the 3 types above).
That way we can provide the list of available scales via IIO, and we can
also convert the scales back to the correct values in the ADC.

Signed-off-by: Alexandru Ardelean <aardelean@baylibre.com>