drivers: power: max42500: Add driver support
The max42500 is a power-system monitor with up to seven voltage monitor
inputs and each input has programmable OV/UV thresholds. The source file
contains the functions needed for the initialization of the device,
reading the registers, writing to the registers, and removing the
device. The header file contains the macros needed for configuration
of the device. Additionally, register files are included.

Signed-off-by: Joshua Maniti <Joshua.Maniti@analog.com>