project:eval-pqmon: small fixes on waveform buffer
Resetting waveform buffer every time processData attribute is set to 0,
this way we can empty for in time use. Also disabled the data copy into
iio buffer in case processData is set to 0, since we empty it every to it
is set.
Changes floating point precision for current and voltage buffer scale.

Signed-off-by: rbudai <robert.budai@analog.com>