add script to convert Windows to Unix paths
When sourcing an environment script such as the Xilinx's settings64.sh,
PATH entries need to be converted to Unix compatible paths. This script
can be used to perform the conversion like this:

source tools/scripts/git-bash-paths.sh /c/Xilinx/Vitis/2023.1/settings64.sh

It's not necessarily specific to Xilinx platform but we're currently
using it only for this purpose with possibility to extend it for other
platforms.

Signed-off-by: Darius Berghe <darius.berghe@analog.com>