tools: generic.mk: add NO_OS_INC_DIRS build var for pulling header files
This is inspired by the SRC_DIRS build variable, but only focuses on
include files. Using the SRC_DIRS can be problematic, as it pulls
every .c & .h file from a directory.

But in some cases, we just want to pull in the header files. These header
files will only get compiled if a .c file pulls them in.

If this build variable is unspecified, it will be empty, so nothing will be
done (to potentially affect current projects).

This needs to be called NO_OS_INC_DIRS, because Mbed already has an
INC_DIRS variable.

Signed-off-by: Alexandru Ardelean <aardelean@baylibre.com>