drivers:meter:ade7978: initial commit
The ADE7978, the ADE7933/ADE7932, and ADE7923 form a
chipset dedicated to measuring 3-phase electrical energy using
shunts as current sensors.

The ADE7933/ADE7932 are isolated, sigma-delta
analog-to-digital converters.
The ADE7923 is a nonisolated, 3-channel Σ-Δ ADC.
ADE7933 has 3 channels and ADE7932 2 channels.

The ADE7978 is a high accuracy, 3-phase electrical energy
measurement IC with serial interfaces and three flexible pulse
outputs.

This is a basic driver implementation with communication done on SPI.

Signed-off-by: Radu Etz <radu.etz@analog.com>