tools:scripts:Set the SPI driver version for Maxim
Since commit [1], the user may choose between two versions of the SPI
driver for the MAX32690 and MAX78002. Currently, we're not defining
this anywhere during the build and as such we get a linker error. Fix
this by selecting v1 as a default value. At this time, the No-OS SPI
drivers for the Maxim platform are based on the v1 version.

The issue first appeared in the v2024_10 release of msdk.

[1] - https://github.com/analogdevicesinc/msdk/commit/755e6ffa0d4627880fcd053435df295e780a742a

Signed-off-by: Ciprian Regus <ciprian.regus@analog.com>