drivers: meter: ade7753: initial commit
The ADE7753 is a Single-Phase Multifunction Metering IC
with di/dt Sensor Interface.
This is a basic driver implementation with communication done over SPI.
The values of the readings are saved in structures in the format read
from the coressponding registers.
Two interrupts are implemented: one for the zero-crossing signal and
another one that serves the irq that can be configured based on the IRQEN
register.

Signed-off-by: Radu Etz <radu.etz@analog.com>