adxl355 - read FIFO as multiples of 3
Data gets stored in the FIFO as locations and 3 locations pertain
to the same measurement (x, y, z). However, it may be that one
interrogates the number of filled locations in the FIFO and this
number is not a multiple of 3 at that moment (most likely due
to locations being filled sequentially). In this case, the correct
approach is to read only multiples of 3 and leave the remaining 1
or 2 locations in the FIFO. This commit does that.

Closes #2431

Signed-off-by: Darius Berghe <darius.berghe@analog.com>