build_projects.yml: Fix Unit Tests job
The ceedling framework used inside the Unit Tests job has released a new
version (>1.0.0) which isn't backwards compatible. This is a hot fix
that sets the version to 0.31.1 (latest working version).
This issue was also reported to the Ceedling repo
(ThrowTheSwitch/Ceedling#939) where porting to the latest versions is
explained.

Signed-off-by: bia1708 <bianca-maria.popu@analog.com>