projects: eval-ade7754: example project Example project for ADE7754. Not tested.
This is a basic example on how to use the ADE7754 driver.
The platform used is Maxim and the MCU board is the AD-APARD32690-SL.
The example uses SPI communication and prints on the serial port
the rms values for all three  phases, the temperature and frequency.
The other measurements are saved in strucutures.

Signed-off-by: Radu Etz <radu.etz@analog.com>