drivers:platform:stm32:stm32_dma.c: Removed support for STM32H563.
The STM32H5 includes a General Purpose DMA as opposed to the Basic DMA in other controllers.
Since there are no way to differentiate BDMA from GPDMA in HAL, controller macros were being
used. This made the code clunky and hard to read. The support for GPDMA will be added in the
successive commits.

Signed-off-by: Swaroop PrakashKumar <Swaroop.PrakashKumar@analog.com>