filter: admv8818: force initialization of SDO
When a weak pull-up is present on the SDO line, spi_update_bits fails
to write both the SOFTRESET and SDOACTIVE bits because it incorrectly
reads them as already set.

Since the soft reset disables the SDO line, performing a
read-modify-write operation on ADI_SPI_CONFIG_A to enable the SDO line
doesn't make sense. This change directly writes to the register instead
of using spi_update_bits.

Fixes: 375520524672 ("drivers:filter: add admv8818 support")
Signed-off-by: Antoniu Miclaus <antoniu.miclaus@analog.com>