doc: sphinx: conf.py: disable smartquotes
The documentation we provide contains also examples of commands which
take input parameters with the double dash prefix (--). Due to the
smartquotes being enabled by default in the sphinx configuration, the
double dash is converted to typographical en-dash (–).

Disable smartquotes such that the commands provided into the
documentation can be used properly without any confusion.

Signed-off-by: Antoniu Miclaus <antoniu.miclaus@analog.com>