fix incremental build from Git Bash
The make that currently is being installed in Git Bash is unable
to look at the timestamp of the files pointed to by symlinks and
compare that to .o timestamps.

Instead, it's comparing the timestamp of the symlink themselves
to the timestamp of .o files.

Because everytime "make all" is invoked, "make update" is called
too, and because "make update" was redoing the symlinks under build/,
a full rebuild was triggered every time.

Fix this by:
 - not removing directories under build/, it's unnecessary
 - keep regenerating symlinks, as before, but make sure to change
   their timestamp to match the one of the source files they point
   to.

The proper make on any linux will ignore the timestamps of the symlinks
so this doesn't break the existing linux workflow. However it improves
the Windows workflow by allowing incremental builds.

It's not an ideal fix, it'd be better to change the tool but at this
point it's harder to change the tool than to do this fix in makefiles.

Signed-off-by: Darius Berghe <darius.berghe@analog.com>