drivers: adc: ad9656: Fixed PN mismatch error
Fixed an issue which would throw a message that said
that the sent PN9 and PN23 sequences don't match. The issue
was present because the chip has two registers which need
to be written with the desired PN sequence in order to test
the PN mechanism. Only one register was written, thus prompting
the mismatch error.

Signed-off-by: Cristian Mihai Popa <cristianmihai.popa@analog.com>