i3c: Remove bus in case no device is attached on the bus
Since the specification foresees hot-join events,
the original thought was that it made sense to keep bus allocated during
the execution life-time, even if no device is attached.
Revisiting it, for a hot-join implementation would make more sense to
have a device attached with a "may_join" flag, and have some sort of
init defer mechanism on the part drivers.
Therefore, harmonize the behaviour with the i2c abstraction of
removing the bus if no device is attached.

Signed-off-by: Jorge Marques <jorge.marques@analog.com>