drivers: adc: ad9208: Added README documentation for AD9208
Signed-off-by: Ramos <cayetonallan.ramos@analog.com>

This adds the .rst file and readme document for AD9208 driver.
Also modified the sphinx driver doc to add the source for this readme
document.

This adds the .rst file and readme document for AD9208 driver.
Also modified the sphinx driver doc to add the source for this readme
document.