projects: eval-adhv4710: example project
Example project for the ADHV4710 evaluation board.
This project is intended to be used with the ADI ADHV4710 evaluation board.
It includes a basic setup for the platform and common data structures.
This example is tested with the AD-APARD32690-SL.
All the alarms are set and enabled in this project
and interogated in the main application file.

Signed-off-by: Radu Etz <radu.etz@analog.com>