drivers: net: oa_tc6: Configurable RX Limit
Add optional CONFIG_OA_THREAD_RX_LIMIT to configure the
rx_limit used by the oa_tc6_thread, which determines the number
of SPI blocks any given call to oa_tc6_thread can do.

For certain applications, especially those with higher throughput,
it may be beneficial for the limit to be lower to allow the caller
to service the ethernet frames to avoid the risk of having no frame buffers
available.

The default value of CONFIG_OA_THREAD_RX_LIMIT is 5 to maintain a default state
which matches the original implementation.

Signed-off-by: Brent Kowal <brent.kowal@analog.com>