drivers:platform:stm32:dma: Fix DMA Abort API
When interrupts are enabled for DMA, the disabling of DMA raises
an interrupt. In the ISR, the flags are to be cleared off.
The DMA_Abort_IT disables the DMA and an interrupt flag is raised.

In no_os_dma.c abort API, IRQ is being disabled and hence
execution won't reach the DMA handler and the flags will never be
cleared. Hence abort needs to be performed without relying on
ISR execution for clearing the flags.

Signed-off-by: Naga Himanshu Indraganti <naga.indraganti@analog.com>